Simplify casting, fix types

+++ lib/json/json.go
@@ -381,7 +381,7 @@ func indent(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, k
 	// it adds a new level of indentation every 2 characters. Clearly, this is
 	// a quadratic growth as the increment grows linearly.
 
-	n := strings.Count(str, "[") + strings.Count(str, "{")
+	n := int64(strings.Count(str, "[") + strings.Count(str, "{"))
 	// Taking into account tabs and newlines and working out the algebra, the
 	// worst case can be compacted in the quadratic formula:
 	worstCase := n*n + 2*n - 1
@@ -398,7 +398,7 @@ func indent(thread *starlark.Thread, b *starlark.Builtin, args starlark.Tuple, k
 	if err := thread.CheckSteps(int64(worstCase)); err != nil {
 		return nil, err
 	}
-	if err := thread.CheckAllocs(int64(len(str) + worstCase*2)); err != nil {
+	if err := thread.CheckAllocs(int64(len(str)) + worstCase*2); err != nil {
 		return nil, err
 	}
 	if err := json.Indent(buf, []byte(str), prefix, indent); err != nil {
